+++ bugsnag-android-performance-impl/src/main/kotlin/com/bugsnag/android/performance/internal/ViewLoadPhase.kt
@@ -1,0 +1,17 @@
+package com.bugsnag.android.performance.internal
+
+import androidx.annotation.RestrictTo
+import com.bugsnag.android.performance.ViewType
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+public enum class ViewLoadPhase(private val phaseName: String) {
+    CREATE("Create"),
+    START("Start"),
+    RESUME("Resume"),
+    DRAW("Draw"),
+    ;
+
+    public fun phaseNameFor(viewType: ViewType): String {
+        return viewType.spanName + phaseName
+    }
+}