+++ bugsnag-plugin-android-performance-appcompat/src/test/kotlin/com/bugsnag/android/performance/FragmentActivityLifecycleCallbacksTest.kt
@@ -3,7 +3,6 @@ package com.bugsnag.android.performance
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import com.bugsnag.android.performance.internal.AutoInstrumentationCache
-import com.bugsnag.android.performance.internal.BugsnagPerformanceInternals
 import com.bugsnag.android.performance.internal.SpanFactory
 import com.bugsnag.android.performance.internal.SpanImpl
 import com.bugsnag.android.performance.internal.SpanTracker
@@ -56,7 +55,7 @@ class FragmentActivityLifecycleCallbacksTest {
 
     @After
     fun shutdown() {
-        BugsnagPerformanceInternals.currentSpanContextStack.clear()
+        SpanContext.defaultStorage?.clear()
     }
 
     /**